Try to target Java 4
Why not x2 ¯\_(ツ)_/¯

I won't go any lower than this for now, as I think I am using a Java 4 feature (logging).

+++ src/main/java/mcTextureGen/generators/AbstractTextureGenerator.java
@@ -34,7 +34,7 @@ public abstract class AbstractTextureGenerator {
         final Random rand;
 
         if (randomSeed != null) {
-            rand = new Random(randomSeed);
+            rand = new Random(randomSeed.longValue());
         } else {
             rand = cachedRand;
         }
